Local analysis

One question that we often want to ask when considering the dynamics of ecological systems is their response to perturbation. One method to answer such questions is to use local stability analysis where one linearises the system around a fixed point and then studies the dynamics of this linear system to learn about the behaviour of the system as a whole.

MiCRM.jl includes a number of tools to carry out such analysis including the automatic calculation of the jacobian matricies which define the linearised system as well as some built in functions to determine key properties of interest for the system.

## Calculating the Jacobian MiCRM.jl automatically calculates the jacobian of a given system using the ForwardDiff.jl package. This provides the automatic-differentiation backend that computes the derivatives of the ODE system at equilibrium. This functionality is all contained within the Analysis.get_jac(sol) function.

MiCRM.Analysis.get_jac — Function

get_jac(sol)

Numerically calculate jacobian of system from the solution object. Assumes that the system is at equilibrium so the end state is used.

Example

TDB

get_jac(sol::SteadyStateSolution; thresh = eps())

For SteadyState solution types

Dynamic Measures

Once the jacobian has been calculated we can use it to calculate several key properties of interest

MiCRM.Analysis.get_Rins — Function

get_Rins(J, t, w = ones(size(J)[1]))

Caclulate the instantaneous rate of growth of the perturbation u at time t. The observation vector w can be optionally supplied to consider linear combinations of the state variables. This can include any linear combination of state variables such as the mass of specific components of the system as well as total functioning measures.

MiCRM.Analysis.get_stability — Function

get_stability(J)

Determine the stability a system given its jaccobian by testing if the real part of the leading eigenvalue is positive.

MiCRM.Analysis.get_reactivity — Function

get_reactivity(J,u)

Test wether the system is "reactive" to the perturbation u. A reactive is system is one where the initial perturbation is amplified so that the deviation from equilibrium initially increases

MiCRM.Analysis.get_return_rate — Function

get_return_rate(J)

get the rate of return of the system from perturbation. This value is determined by the leading eigenvalue regardless of the direction of the perturbation or the observation vector we choose.