TBD: Contains info on the specific parameter generating functions included already

MiCRM.Parameters.default_m — Method

default params

MiCRM.Parameters.generate_params — Method

generate_params(N,M;f_m = default_m, f_ρ = default_ρ, 
    f_ω = default_ω, f_u = default_u, 
    f_l = default_l, kwargs...)

Function to generate parameter sets for MiCRM simualtions. Requires the system size is defined as well as the functions to generate the actual parameters used in the simulations. Extra arguments can also be passed via the kwargs argument.

By default the function generates a random MiCRM system with no structure and uptake and leakage matricies generated by Dirichlet distribitions.

The function returns a NamedTuple with all the parameters as well as a kw entry which itself is a NamedTuple with the additional arguments provided.

MiCRM.Parameters.modular_leakage — Method

modular_leakage(N,M,kw::Dict{Symbol, Any})

Wrapper for the modular leakage to allow it to be used in the generate_params function.

MiCRM.Parameters.modular_leakage — Method

modular_leakage(M; N_modules = 2, s_ratio = 10.0, λ = 0.5)

Function that generates a modular leakage matrix for a system with M resources with a directional leakage structure. The matrix is generated using a Dirichlet distribution such that the leakage of all resources sums to λ.

The number of modules determines how many groups of resources they are split into. For example if N_modules = 5 then the resources will be split into five groups with the first group of resources tending to leak to the second, the second to the third and so on.

The degree of to which resources leak in this constrained way (verses randomly across all resources) is determined by the s_ratio value. This controls the relative value of the dirchlet α parameters which determine how the probabiltiy density function is distributed over the different resources. A greater probabiltiy of leakage from one resource to another is obtained by setting the α values for resources to higher values. When s_ratio = 1 the proabailtiy is uniform and all resources are equally likely to be leaked to each other. When s_ratio >1 then resources are more likely to leak to resources thier own or the next module in the sequence.

MiCRM.Parameters.modular_uptake — Method

modular_uptake(N,M,kw)

Wrapper for the modular uptake to allow it to be used in the generate_params function.

MiCRM.Parameters.modular_uptake — Method

modular_uptake(N,M; N_modules = 2, s_ratio = 10.0)

Function that generates a modular uptake matrix for a system with N consumers and M resources. The matrix is generated using a Dirichlet distribution such that the uptake of all consumers sums to 1.

The number of modules determines how many groups of resources the consumers are specialised over. For example if N_modules = 2 then the resources will be split into two groups with half the consumers specialising on one and half on the other.

The degree of specialisation is determined by the s_ratio value. This controls the relative value of the dirchlet α parameters which determine how the probabiltiy density function is distributed over the different resources. Specialisation is obtained by setting the α values for resources that specialists consume to higher values meaning they have a higher probablity of a larger value. When s_ratio = 1 the proabailtiy is uniform and all resources are equally likely to be consumed. When s_ratio >1 then consumers are more likely to consume resources within thier module.